requirements.txt
pandas>=2.0.0
numpy>=1.24.0
faker>=18.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
sqlalchemy>=2.0.0
streamlit>=1.25.0
plotly>=5.16.0
psycopg2-binary>=2.9.0